Implement dropout-based data availability for lightweight tasks
When a lightweight task runs, if it has a data availability policy, the client will mark it as eligible by default. Then, when the task executes, the example_query_plan_engine will make sure that there are at least min_output_row_count results in the output. If there are not, the device will drop out.

This adds a new TrainOutcome and PhaseOutcome indicating insufficient data. We use the FAILED_PRECONDITON status code to indicate this to the server.

PiperOrigin-RevId: 823079977