Add metadata and metadata config protos.
We'll use MetadataContainerConfig in the access policy's config constraints for the metadata container. The metadata container will create a PayloadMetadataSet which contains metadata, each identified by name. Downstream pipelines that use the metadata can select their desired metadata using this name.

We define a separate TeePayloadMetadata since the existing PayloadMetadata is still used by the client and it should be clear that the new TEE-generated metadata isn't part of the uploaded blob header from the client. Eventually when we get rid of client event time range we can deprecate the original PayloadMetadata message.

Also add an initialization config proto for the metadata container. For PrivateLogger uploads we need the customer logged proto's name and FileDescriptorSet so we can parse the logged protos to Messages and use the refleciton API.

PiperOrigin-RevId: 832458250